requirements.txt
asgiref==3.7.2
Django
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
nanoid==2.0.0
PyJWT==2.8.0
pytz==2024.1
setuptools==69.1.0
sqlparse==0.4.4
mysqlclient
django-templated-mail